ShoppingCart BugFix

+++ BookStore.Core/ViewModels/ShoppingCartVM.cs
@@ -10,7 +10,7 @@ namespace BookStore.Core.ViewModels
     public class ShoppingCartVM
     {
         public IEnumerable<ShoppingCart> ShoppingCartList { get; set; }
-        public decimal OrderTotal { get; set; }
+        public double OrderTotal { get; set; }
 
     }
 }
